apps/api/pyproject.toml
[project]
name = "re-api"
version = "0.1.0"
description = "Revenue Edge API Gateway"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.5.0",
  "supabase>=2.8.0",
  "httpx>=0.27.0",
  "pyjwt>=2.9.0",
  "python-multipart>=0.0.12",
  "sentry-sdk[fastapi]>=2.14.0",
  "twilio>=9.3.0",
  "retell-sdk>=4.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
  "black>=24.8.0",
  "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
